fi: seascanarch changes for 64-bit

+++ include/wintypes.h
@@ -9,6 +9,9 @@
 #define MS_STRUCT_FILLER(X, A) 
 
 #else
+
+#include <stdint.h>
+
 // Define filler elements so the alignment of structs works correctly
 // on unix.
 // FIXME:  how does this work in MINGW?  Seems I've tried every possible
@@ -16,24 +19,24 @@
 
 #define MS_STRUCT __attribute__((packed))
 #define MS_STRUCT_FILLER(NAME, BYTES) char __filler__##NAME [BYTES]
-
+//#define MS_STRUCT_FILLER(NAME, BYTES)
 // typedefs needed for IDLCONST.H and IDLTYPES.H
 
-typedef long LONG;
-typedef unsigned long ULONG;
-typedef unsigned short USHORT;
+typedef int32_t LONG;
+typedef uint32_t ULONG;
+typedef uint16_t USHORT;
 typedef unsigned char BYTE;
-typedef int BOOL;
-typedef short SHORT;
-typedef unsigned short WORD;
+typedef int32_t BOOL;
+typedef int16_t SHORT;
+typedef uint16_t WORD;
 typedef unsigned char UCHAR;
 typedef char CHAR;
 typedef void * HANDLE;
-typedef long long LONGLONG;
-typedef unsigned long long ULONGLONG;
+typedef int64_t LONGLONG;
+typedef uint64_t ULONGLONG;
 typedef float FLOAT;
 typedef double DOUBLE;
-typedef unsigned int DWORD;
+typedef uint32_t DWORD;
 #define __stdcall 
 
 typedef union _LARGE_INTEGER {
